Use BigDecimal instead of Double
Replaced the Double for class of currencies "pricing" and "managementFee" with BigDecimal

+++ src/main/java/com/project/homes/entity/Apartment.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.Objects;
+import java.math.BigDecimal;
+
 
 @Data
 @Document(collection = "Apartments")
@@ -11,11 +13,11 @@ public class Apartment extends Property{
     private String unit;
     private Boolean hasParkingLot;
     private Boolean hasGym;
-    private Double managementFee;
+    private BigDecimal managementFee;
 
-    public Apartment(PropertyType type, Double area, Double pricing, Address address, int numOfRooms,
+    public Apartment(PropertyType type, Double area, BigDecimal pricing, Address address, int numOfRooms,
                      int numOfBaths, String pictureUrl, String ownerId, String unit, boolean hasParkingLot,
-                     boolean hasGym, Double managementFee)
+                     boolean hasGym, BigDecimal managementFee)
     {
         super(type, area, pricing, address, numOfRooms, numOfBaths, pictureUrl, ownerId);
         this.unit = unit;
